workspace/RAG/rag_api/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pydantic>=2.0.2,<3.0.0
pyyaml==6.0
requests==2.31.0
